Make some ResponseStateOperations public
