+++ include/ad_rss/state/ResponseStateOperation.hpp
@@ -1,0 +1,85 @@
+// ----------------- BEGIN LICENSE BLOCK ---------------------------------
+//
+// Copyright (c) 2018-2019 Intel Corporation
+//
+// Redistribution and use in source and binary forms, with or without modification,
+// are permitted provided that the following conditions are met:
+//
+// 1. Redistributions of source code must retain the above copyright notice,
+//    this list of conditions and the following disclaimer.
+//
+// 2. Redistributions in binary form must reproduce the above copyright notice,
+//    this list of conditions and the following disclaimer in the documentation
+//    and/or other materials provided with the distribution.
+//
+// 3. Neither the name of the copyright holder nor the names of its contributors
+//    may be used to endorse or promote products derived from this software without
+//    specific prior written permission.
+//
+//    THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+//    ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+//    WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED.
+//    IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
+//    INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+//    BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA,
+//    OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
+//    WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+//    ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+//    POSSIBILITY OF SUCH DAMAGE.
+//
+// ----------------- END LICENSE BLOCK -----------------------------------
+/**
+ * @file
+ */
+
+#pragma once
+
+#include "ad_rss/state/ResponseState.hpp"
+
+/*!
+ * @brief namespace ad_rss
+ */
+namespace ad_rss {
+/*!
+ * @brief namespace for RSS state datatypes and operations
+ */
+namespace state {
+
+/**
+ * @brief is response state longitudinal safe
+ *
+ * @param[in] responseState to check
+ *
+ * true if safe, false if not
+ */
+inline bool isLongitudinalSafe(ResponseState const &responseState)
+{
+  return responseState.longitudinalState.isSafe;
+}
+
+/**
+ * @brief is response state lateral safe
+ *
+ * @param[in] responseState to check
+ *
+ * true if safe, false if not
+ */
+inline bool isLateralSafe(ResponseState const &responseState)
+{
+  return responseState.lateralStateRight.isSafe && responseState.lateralStateLeft.isSafe;
+}
+
+/**
+ * @brief is response state dangerous
+ *
+ * @param[in] responseState to check
+ *
+ * true if dangerous, false if not
+ */
+inline bool isDangerous(ResponseState const &responseState)
+{
+  return !isLongitudinalSafe(responseState) && !isLateralSafe(responseState);
+}
+
+} // namespace state
+} // namespace ad_rss
